+++ data_utils/datahub_services/__init__.py
@@ -1,0 +1,23 @@
+"""
+DataHub 유틸리티 패키지
+
+DataHub와의 상호작용을 위한 모듈들을 제공합니다.
+
+주요 구성요소:
+- DataHubBaseClient: 기본 연결 및 통신
+- MetadataService: 메타데이터, 리니지, URN 관련 기능
+- QueryService: 쿼리 관련 기능  
+- GlossaryService: 용어집 관련 기능
+"""
+
+from .base_client import DataHubBaseClient
+from .metadata_service import MetadataService
+from .query_service import QueryService
+from .glossary_service import GlossaryService
+
+__all__ = [
+    'DataHubBaseClient',
+    'MetadataService', 
+    'QueryService',
+    'GlossaryService'
+]
